+++ packages/core/client/Router.js
@@ -358,15 +358,16 @@ Router.map(function() {
 		layoutTemplate: 'layoutMap',
 		waitOn: function() {
 			Session.set('showSidebar', false);
-			return Meteor.subscribe('placeGeometryById', this.params.placeId);
+			return Meteor.subscribe('placeById', this.params.placeId);
 		},
 		onAfterAction: function() {
 			if(!this.ready()) return null;
 
 			var place = K.placeById( this.params.placeId );
 
-			if(place)
+			if(place) {
 				place.update().showGeometry();
+			}
 		}
 	});